// src/generated/graphql.tsx
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: any;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
  /** The `Long` scalar type represents 52-bit integers */
  Long: any;
  /** The `Upload` scalar type represents a file upload. */
  Upload: any;
};
// ...
export type Cart = {
  __typename?: 'Cart';
  cartItems?: Maybe<Array<Maybe<ComponentProductCartItem>>>;
  createdAt?: Maybe<Scalars['DateTime']>;
  total?: Maybe<Scalars['Float']>;
  updatedAt?: Maybe<Scalars['DateTime']>;
};


export type CartCartItemsArgs = {
  filters?: InputMaybe<ComponentProductCartItemFiltersInput>;
  pagination?: InputMaybe<PaginationArg>;
  sort?: InputMaybe<Array<InputMaybe<Scalars['String']>>>;
};

export type CartEntity = {
  __typename?: 'CartEntity';
  attributes?: Maybe<Cart>;
  id?: Maybe<Scalars['ID']>;
};

export type CartEntityResponse = {
  __typename?: 'CartEntityResponse';
  data?: Maybe<CartEntity>;
};
// ...
export type CategoryRelationResponseCollection = {
  __typename?: 'CategoryRelationResponseCollection';
  data: Array<CategoryEntity>;
};

export type ComponentProductCartItem = {
  __typename?: 'ComponentProductCartItem';
  id: Scalars['ID'];
  image?: Maybe<UploadFileEntityResponse>;
  name: Scalars['String'];
  price: Scalars['Float'];
  product?: Maybe<ProductEntityResponse>;
  quantity: Scalars['Int'];
  variant?: Maybe<VariantEntityResponse>;
};
// ...
export type ComponentProductCartItemInput = {
  id?: InputMaybe<Scalars['ID']>;
  image?: InputMaybe<Scalars['ID']>;
  name?: InputMaybe<Scalars['String']>;
  price?: InputMaybe<Scalars['Float']>;
  product?: InputMaybe<Scalars['ID']>;
  quantity?: InputMaybe<Scalars['Int']>;
  variant?: InputMaybe<Scalars['ID']>;
};

export type ComponentProductColor = {
  __typename?: 'ComponentProductColor';
  hex?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  name: Scalars['String'];
};

export type ComponentProductOption = {
  __typename?: 'ComponentProductOption';
  id: Scalars['ID'];
  name: Scalars['String'];
  values: Array<Maybe<ComponentProductValue>>;
};
// ...
export type MutationAddCartItemArgs = {
  cartItem: ComponentProductCartItemInput;
  id: Scalars['ID'];
};
// ...
export type MutationDeleteCartItemArgs = {
  cartItemId: Scalars['ID'];
  id: Scalars['ID'];
};
// ...
export type MutationEmptyCartArgs = {
  id: Scalars['ID'];
};
// ...
export type MutationUpdateCartItemArgs = {
  cartItemId: Scalars['ID'];
  id: Scalars['ID'];
  quantity: Scalars['Int'];
};
// ...
export type Product = {
  __typename?: 'Product';
  categories?: Maybe<CategoryRelationResponseCollection>;
  createdAt?: Maybe<Scalars['DateTime']>;
  description: Scalars['String'];
  image: UploadFileEntityResponse;
  name: Scalars['String'];
  options?: Maybe<Array<Maybe<ComponentProductOption>>>;
  price: Scalars['Float'];
  publishedAt?: Maybe<Scalars['DateTime']>;
  updatedAt?: Maybe<Scalars['DateTime']>;
  variants?: Maybe<VariantRelationResponseCollection>;
};
// ...
export type ProductEntity = {
  __typename?: 'ProductEntity';
  attributes?: Maybe<Product>;
  id?: Maybe<Scalars['ID']>;
};

export type ProductEntityResponse = {
  __typename?: 'ProductEntityResponse';
  data?: Maybe<ProductEntity>;
};
// ...
export type UploadFileEntityResponse = {
  __typename?: 'UploadFileEntityResponse';
  data?: Maybe<UploadFileEntity>;
};
// ...
export type Variant = {
  __typename?: 'Variant';
  createdAt?: Maybe<Scalars['DateTime']>;
  description?: Maybe<Scalars['String']>;
  image?: Maybe<UploadFileEntityResponse>;
  name: Scalars['String'];
  price: Scalars['Float'];
  product?: Maybe<ProductEntityResponse>;
  selectedOptions: Scalars['JSON'];
  updatedAt?: Maybe<Scalars['DateTime']>;
};

export type VariantEntity = {
  __typename?: 'VariantEntity';
  attributes?: Maybe<Variant>;
  id?: Maybe<Scalars['ID']>;
};
// ...
export type VariantRelationResponseCollection = {
  __typename?: 'VariantRelationResponseCollection';
  data: Array<VariantEntity>;
};